// src/store/action-types/banner-action-type/banner-action.type.ts
export type InitialBannerState = {
  data?: any[];
  page?: number;
  limit?: number;
  total?: number;
  sortBy?: string;
  sortType?: 'asc' | 'desc';
  search?: string;
  groupId?: string;
  categoryId?: string;
  productCategoryList?: any[];
  productGroupList?: any[];
  billerList?: any[];
  productCategoryFormList?: any[];
  productGroupFormList?: any[];
  loadingList?: boolean;
  loadingPost?: boolean;
  loadingDetail?: boolean;
  loadingDelete?: boolean;
  loading?: boolean;
  error?: null | string;
};

export type Banner = {
  type: string;
  payload: InitialBannerState;
};

export type BannerDispatchType = (args: Banner) => Banner;